refactor(dashboard | login): rename filenames to kebab-case and organize component folder for context related features
