--- src/app/dashboard/_components/listView/index.tsx
@@ -1,1 +1,0 @@
-export { ListView } from "./listView"
