+++ src/components/dashboard/list-view/index.tsx
@@ -1,0 +1,1 @@
+export { ListView } from "./list-view"
